Add to README.md

+++ src/main/resources/db/migration/20170306_1611__Initial_tables.sql
@@ -33,14 +33,14 @@ CREATE TABLE data
     high DOUBLE,
     low DOUBLE,
     close DOUBLE,
-    volume INT(15),
+    volume BIGINT(20),
     ex_dividend DOUBLE,
     split_ratio DOUBLE,
     adj_open DOUBLE,
     adj_high DOUBLE,
     adj_low DOUBLE,
     adj_close DOUBLE,
-    adj_volume INT(15)
+    adj_volume BIGINT(20)
 );
 CREATE INDEX data_date_index ON data (date);
 CREATE UNIQUE INDEX data_id_uindex ON data (id);
